update total bag
